マイナスと小数があり得るのでuintではダメだった

+++ src/rendering/shader/compute.wgsl
@@ -12,21 +12,21 @@ struct Uniforms {
 }
 
 struct IterationInputMeta {
-  x: u32,
-  y: u32,
-  width: u32,
-  height: u32,
-  resolutionWidth: u32,
-  resolutionHeight: u32,
-  length: u32, // element length (not byte)
-  isSuperSampled: u32, // 0 or 1
+  x: f32,
+  y: f32,
+  width: f32,
+  height: f32,
+  resolutionWidth: f32,
+  resolutionHeight: f32,
+  length: f32, // element length (not byte)
+  isSuperSampled: f32, // 0 or 1
 };
 
 @group(0) @binding(0) var<uniform> uniforms: Uniforms;
 @group(0) @binding(1) var<storage, read_write> iterations: array<u32>;
 // @group(0) @binding(2) var<storage> palette: array<vec4f>;
 @group(0) @binding(3) var<storage> iterInput: array<u32>;
-@group(0) @binding(4) var<storage> iterInputMeta: array<u32>;
+@group(0) @binding(4) var<storage> iterInputMeta: array<f32>;
 
 fn isValidIdx(idx: i32, length: i32) -> bool {
   return idx >= 0 && idx < length;
